Add files via upload
This is the data for six test games. Each move has: the number of times this move was made by the losing side (-), the number of times this move was made by the winning side (+), the percentage of times this move won out of all times the move was made.
Board spaces are numbered from 1 at the top left to 32 at the bottom right. Only spaces where checkers can be placed are numbered.